FEATURE OrderFactory - record purchases for products, orders
move the load of calculating inventory for products and variations from the add to cart form to order processing.

As each order is processed, it will query all OrderDetail records and record the number of purchases for product records. It will also query all OrderVariation records and record the number of purchases for Variations